LibWeb: Store text chunks in the rendered text snapshot
Text chunks lived in a second arena table with their own generation
checks and cleanup. Publishing rendered text had to explicitly
invalidate that table even though the chunks were derived from the text
snapshot.

Keep the chunk cache inside TextContent beside its grapheme segmenter.
Snapshot replacement and node removal now release the cached entry
automatically, while identical publication preserves it. Retain shared
ownership of chunk entries and their fonts so active iterators survive
nested measurements and snapshot replacement.

Extend the chunk lifetime test to cover identical publication, text
replacement, and node removal in addition to replacement of the chunk
cache itself.
